+++ app/build.gradle
@@ -76,6 +76,7 @@ dependencies {
     compile 'com.ncapdevi:frag-nav:1.0.3'
     compile 'com.makeramen:roundedimageview:2.2.1'
     compile 'com.github.arimorty:floatingsearchview:2.0.3'
+    compile 'com.github.paolorotolo:appintro:4.1.0'
 }
 
 apply plugin: 'com.google.gms.google-services'
