+++ app/src/main/res/drawable/background_splash.xml
@@ -10,7 +10,7 @@
     </item>
 
     <item>
-        <bitmap android:src="@drawable/start_p0_circle_binners"
+        <bitmap android:src="@drawable/binners_logo_negative"
             android:gravity="center"/>
 
     </item>
