--- app/src/main/res/layout/about_app_layout.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v4.view.ViewPager xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/my_view_pager"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-</android.support.v4.view.ViewPager>
