+++ app/src/main/res/layout/onboarding_1_fragment.xml
@@ -15,7 +15,6 @@
     android:layout_width="match_parent"
     android:layout_height="100dp"
     tools:background="@color/blue"
-    android:layout_marginTop="48dp"
     >
     <ca.com.androidbinnersproject.customViews.BinnersTextView
       android:layout_width="match_parent"
@@ -81,7 +80,7 @@
     android:layout_gravity="center"
     android:orientation="vertical"
     tools:background="@color/red"
-    android:layout_marginBottom="48dp"
+    android:layout_marginBottom="64dp"
     >
 
     <ca.com.androidbinnersproject.customViews.BinnersTextView
@@ -94,7 +93,6 @@
       android:textColor="@color/white"
       android:textSize="16sp"
       app:btv_textStyle="regular"
-      android:lines="5"
       />
 
   </FrameLayout>
