+++ app/src/main/res/layout/onboarding_3_fragment.xml
@@ -14,7 +14,6 @@
   <FrameLayout
     android:layout_width="match_parent"
     android:layout_height="100dp"
-    android:layout_marginTop="48dp"
     tools:background="@color/blue"
     >
 
@@ -81,7 +80,7 @@
   <LinearLayout
     android:layout_width="match_parent"
     android:layout_height="100dp"
-    android:layout_marginBottom="48dp"
+    android:layout_marginBottom="64dp"
     android:orientation="vertical"
     tools:background="@color/red"
     >
